2/23/24 Create infinitemonkeys.cpp
C++ version of code that takes a string as input and randomly generates strings of that length until it hits that word. It's named after the so called "infinite monkey theorem" stating that monkeys typing on a keyboard for long enough would eventually type out all the works of Shakespeare.